+++ src/ray/core_worker/BUILD.bazel
@@ -14,7 +14,7 @@ ray_cc_library(
     deps = [
         ":actor_handle",
         ":actor_manager",
-        ":core_worker_common",
+        ":common",
         ":core_worker_context",
         ":core_worker_options",
         ":experimental_mutable_object_manager",
@@ -62,7 +62,7 @@ ray_cc_library(
     name = "core_worker_options",
     hdrs = ["core_worker_options.h"],
     deps = [
-        ":core_worker_common",
+        ":common",
         "//src/ray/common:id",
         "//src/ray/common:ray_object",
         "//src/ray/common:status",
@@ -92,18 +92,7 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "scheduling_queue",
-    hdrs = ["transport/scheduling_queue.h"],
-    deps = [
-        "//src/ray/common:id",
-        "//src/ray/common:task_common",
-        "//src/ray/protobuf:core_worker_cc_proto",
-        "//src/ray/rpc:server_call",
-    ],
-)
-
-ray_cc_library(
-    name = "core_worker_common",
+    name = "common",
     srcs = ["common.cc"],
     hdrs = ["common.h"],
     deps = [
@@ -112,6 +101,7 @@ ray_cc_library(
         "//src/ray/common:task_common",
         "//src/ray/raylet_client:raylet_client_lib",
     ],
+    visibility = [":__subpackages__"],
 )
 
 ray_cc_library(
@@ -119,7 +109,7 @@ ray_cc_library(
     srcs = ["context.cc"],
     hdrs = ["context.h"],
     deps = [
-        ":core_worker_common",
+        ":common",
         "//src/ray/common:runtime_env",
         "//src/ray/common:task_common",
         "@boost//:thread",
@@ -161,10 +151,9 @@ ray_cc_library(
         ":actor_creator",
         ":actor_handle",
         ":actor_task_submitter",
-        ":core_worker_common",
+        ":common",
         ":core_worker_context",
         ":reference_count",
-        ":task_receiver",
         "//src/ray/common:id",
         "//src/ray/common:task_common",
         "//src/ray/gcs:gcs_pb_util",
@@ -227,6 +216,7 @@ ray_cc_library(
         "@com_google_absl//absl/synchronization",
         "@com_google_absl//absl/types:optional",
     ],
+    visibility = [":__subpackages__"],
 )
 
 ray_cc_library(
@@ -346,121 +336,6 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "scheduling_util",
-    srcs = ["transport/scheduling_util.cc"],
-    hdrs = ["transport/scheduling_util.h"],
-    deps = [
-        "//src/ray/common:id",
-        "//src/ray/common:task_common",
-        "//src/ray/protobuf:core_worker_cc_proto",
-        "//src/ray/raylet_client:raylet_client_lib",
-        "//src/ray/rpc:server_call",
-    ],
-)
-
-ray_cc_library(
-    name = "normal_scheduling_queue",
-    srcs = ["transport/normal_scheduling_queue.cc"],
-    hdrs = ["transport/normal_scheduling_queue.h"],
-    deps = [
-        "scheduling_queue",
-        ":scheduling_util",
-        "//src/ray/common:id",
-        "//src/ray/common:task_common",
-        "//src/ray/rpc:server_call",
-        "@com_google_absl//absl/base:core_headers",
-        "@com_google_absl//absl/synchronization",
-    ],
-)
-
-ray_cc_library(
-    name = "actor_scheduling_queue",
-    srcs = ["transport/actor_scheduling_queue.cc"],
-    hdrs = ["transport/actor_scheduling_queue.h"],
-    deps = [
-        ":concurrency_group_manager",
-        ":core_worker_fiber",
-        ":scheduling_queue",
-        ":scheduling_util",
-        ":task_event_buffer",
-        ":thread_pool",
-        "//src/ray/common:id",
-        "//src/ray/common:task_common",
-        "//src/ray/protobuf:core_worker_cc_proto",
-        "//src/ray/raylet_client:raylet_client_lib",
-        "//src/ray/rpc:server_call",
-        "@com_google_absl//absl/base:core_headers",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/container:flat_hash_set",
-        "@com_google_absl//absl/synchronization",
-    ],
-)
-
-ray_cc_library(
-    name = "concurrency_group_manager",
-    srcs = ["transport/concurrency_group_manager.cc"],
-    hdrs = ["transport/concurrency_group_manager.h"],
-    deps = [
-        ":core_worker_fiber",
-        ":thread_pool",
-        "//src/ray/common:task_common",
-    ],
-)
-
-ray_cc_library(
-    name = "out_of_order_actor_scheduling_queue",
-    srcs = ["transport/out_of_order_actor_scheduling_queue.cc"],
-    hdrs = ["transport/out_of_order_actor_scheduling_queue.h"],
-    deps = [
-        ":concurrency_group_manager",
-        ":core_worker_fiber",
-        ":scheduling_queue",
-        ":scheduling_util",
-        ":task_event_buffer",
-        ":thread_pool",
-        "//src/ray/common:id",
-        "//src/ray/common:task_common",
-        "//src/ray/protobuf:core_worker_cc_proto",
-        "//src/ray/raylet_client:raylet_client_lib",
-        "//src/ray/rpc:server_call",
-        "@com_google_absl//absl/base:core_headers",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/container:flat_hash_set",
-        "@com_google_absl//absl/synchronization",
-    ],
-)
-
-ray_cc_library(
-    name = "task_receiver",
-    srcs = ["transport/task_receiver.cc"],
-    hdrs = ["transport/task_receiver.h"],
-    deps = [
-        ":actor_creator",
-        ":actor_handle",
-        ":actor_scheduling_queue",
-        ":actor_task_submitter",
-        ":concurrency_group_manager",
-        ":core_worker_common",
-        ":core_worker_context",
-        ":core_worker_fiber",
-        ":dependency_resolver",
-        ":memory_store",
-        ":normal_scheduling_queue",
-        ":out_of_order_actor_scheduling_queue",
-        ":thread_pool",
-        "//src/ray/common:asio",
-        "//src/ray/common:id",
-        "//src/ray/common:ray_object",
-        "//src/ray/common:task_common",
-        "//src/ray/rpc:server_call",
-        "@com_google_absl//absl/base:core_headers",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/container:flat_hash_set",
-        "@com_google_absl//absl/synchronization",
-    ],
-)
-
-ray_cc_library(
     name = "experimental_mutable_object_manager",
     srcs = ["experimental_mutable_object_manager.cc"],
     hdrs = ["experimental_mutable_object_manager.h"],
@@ -505,7 +380,7 @@ ray_cc_library(
     srcs = ["generator_waiter.cc"],
     hdrs = ["generator_waiter.h"],
     deps = [
-        ":core_worker_common",
+        ":common",
         "@com_google_absl//absl/synchronization",
     ],
 )
@@ -541,7 +416,7 @@ ray_cc_library(
     srcs = ["store_provider/plasma_store_provider.cc"],
     hdrs = ["store_provider/plasma_store_provider.h"],
     deps = [
-        ":core_worker_common",
+        ":common",
         ":core_worker_context",
         ":reference_count",
         "//src/ray/common:buffer",
@@ -567,22 +442,10 @@ ray_cc_library(
         ":lease_policy",
         ":memory_store",
         ":task_manager",
-        ":task_receiver",
         "//src/ray/common:id",
         "//src/ray/gcs:gcs_pb_util",
         "//src/ray/raylet_client:raylet_client_lib",
         "//src/ray/rpc:core_worker_client",
         "@com_google_absl//absl/base:core_headers",
-    ],
-)
-
-ray_cc_library(
-    name = "thread_pool",
-    srcs = ["transport/thread_pool.cc"],
-    hdrs = ["transport/thread_pool.h"],
-    deps = [
-        "//src/ray/util:logging",
-        "@boost//:asio",
-        "@boost//:thread",
     ],
 )
