+++ src/ray/core_worker/test/BUILD.bazel
@@ -29,23 +29,6 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "task_receiver_test",
-    srcs = ["task_receiver_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//:ray_mock",
-        "//src/ray/common:asio",
-        "//src/ray/common:task_common",
-        "//src/ray/common:test_util",
-        "//src/ray/core_worker:normal_task_submitter",
-        "//src/ray/core_worker:reference_count",
-        "//src/ray/rpc:core_worker_client",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
     name = "actor_task_submitter_test",
     srcs = ["actor_task_submitter_test.cc"],
     tags = ["team:core"],
@@ -155,59 +138,6 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "scheduling_queue_test",
-    srcs = ["scheduling_queue_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//src/ray/common:asio",
-        "//src/ray/common:test_util",
-        "//src/ray/core_worker:actor_scheduling_queue",
-        "//src/ray/core_worker:core_worker_lib",
-        "//src/ray/core_worker:scheduling_queue",
-        "//src/ray/core_worker:task_receiver",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "thread_pool_test",
-    srcs = ["thread_pool_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//src/ray/core_worker:thread_pool",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "concurrency_group_manager_test",
-    srcs = ["concurrency_group_manager_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//src/ray/common:asio",
-        "//src/ray/common:test_util",
-        "//src/ray/core_worker:concurrency_group_manager",
-        "//src/ray/core_worker:task_receiver",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "fiber_state_test",
-    srcs = ["fiber_state_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//src/ray/core_worker:core_worker_fiber",
-        "//src/ray/util:logging",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
     name = "actor_submit_queue_test",
     size = "small",
     srcs = ["actor_submit_queue_test.cc"],
@@ -309,7 +239,7 @@ ray_cc_test(
     deps = [
         "//:ray_mock",
         "//src/ray/common:test_util",
-        "//src/ray/core_worker:core_worker_common",
+        "//src/ray/core_worker:common",
         "//src/ray/core_worker:generator_waiter",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "@com_google_googletest//:gtest",
